Last 1.4 code update
This is the last working version which still uses React.NET. As of writing, I had begun working on a new version that relies on a single library, the one I forked so you'll always have at least old binaries. Thus I apologise but I cannot remember what changes were made to this version.

If you are fine with React.NET, if you haven't run into the environment errors, then I recommend this to you.