FEngine, provides RenderHtml and TransformCode
FEngine is my replacement for React.NET and responds in about a third the time, the first call of RenderHtml is dreadfully slow and the second isn't much better but I think this is due to some optimisation causing a static object to be instanstiated late as well as some internal VroomJs caching not kicking in early enough. Otherwise fairly unchanging props or consistent and expected props will give faster response times than React.NET.

The other added bonus is that now this project is truly cross platform as the only library React.NET was accommodating for which was cross platform was VroomJs, as each other was specifically windows. For whatever reason, React.NET wouldn't always load the native VroomJs binary, but when using VroomJs directly the binary could be loaded in the same directory as the debug or release folder.